requirements.txt
<<<<<<< HEAD
discord.py>=2.6.0
yt-dlp>=2025.4.25
python-dotenv>=1.0.0
PyNaCl>=1.5.0

=======
# Discord Bot 기본 환경 (최소 버전만 지정)
discord.py>=2.6.0
PyNaCl>=1.6.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0

# 의존성 (aiohttp 등 비슷한 최신 버전과 호환)
aiohttp>=3.8.0
aiohappyeyeballs>=2.0.0
aiosignal>=1.3.0
attrs>=23.0.0
audioop-lts>=0.2.0
cffi>=1.15.0
frozenlist>=1.3.0
idna>=3.4
multidict>=6.0.0
propcache>=0.4.0
pycparser>=2.21
yarl>=1.9.0
>>>>>>> master
